dbft: move n3adaptors to consensus/dbft package
Don't need to use a separate folder for them, these N3 adaptors are
closely related to the consensus. It also allows to avoid cyclic dependencies
in future.
